connecting view and database
