+++ app/Http/Controllers/TaskController.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Models\Task;
+use Illuminate\Http\Request;
+
+class TaskController extends Controller
+{
+    public function index()
+    {
+        $Tasks = Task::query()->get();
+        return view('myTasks.index', compact('Tasks'));
+    }
+
+    public function create()
+    {
+        return view('myTasks.create');
+    }
+
+    public function store(Request $request)
+    {
+        //
+    }
+
+
+    public function show(string $id)
+    {
+        //
+    }
+
+
+    public function edit(string $id)
+    {
+        //
+    }
+
+
+    public function update(Request $request, string $id)
+    {
+        //
+    }
+
+    public function destroy(string $id)
+    {
+        //
+    }
+
+}
