+++ resources/views/myCategories/index.blade.php
@@ -1,152 +1,11 @@
-<style>
-    .myCategory-box {
-        background: white;
-        padding: 1.5rem;
-        border-radius: 12px;
-        box-shadow: var(--card-shadow);
-        height: auto;
-    }
+@extends('layouts.master')
 
-    .btn {
-        padding: 10px 20px;
-        border: none;
-        border-radius: 5px;
-        cursor: pointer;
-        transition: all 0.3s ease;
-        font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    }
+@section('head')
+    <title>myCats</title>
+    <link rel="stylesheet" href="{{asset('app/myCats/categoryStyle.css')}}">
+@endsection
 
-    .row {
-        background: white;
-        padding: 1.5rem;
-        border-radius: 12px;
-        box-shadow: var(--card-shadow);
-    }
-
-    .Category-table {
-        width: 100%;
-        border-collapse: collapse;
-        margin: 10px 0;
-        font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-        background: #fff;
-        color: #000;
-    }
-
-    .Category-table th,
-    .Category-table td {
-        padding: 12px 15px;
-        text-align: right;
-    }
-
-    .Category-table th {
-        font-weight: 600;
-    }
-
-    .action-btn {
-        padding: 5px 10px;
-        margin: 0 3px;
-        border: none;
-        border-radius: 3px;
-        cursor: pointer;
-        transition: all 0.3s;
-    }
-
-    .edit-btn {
-        background-color: #666;
-        color: white;
-        box-shadow: 0 2px 8px rgba(0, 0, 0, 0.5);
-        padding: 11px;
-        border-radius: 14px;
-    }
-
-    .delete-btn {
-        background-color: red;
-        color: white;
-        box-shadow: 0 2px 8px rgba(0, 0, 0, 0.5);
-        padding: 11px;
-        border-radius: 14px;
-    }
-
-    .action-btn:hover {
-        opacity: 0.6;
-    }
-    .modal-overlay {
-        position: fixed;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-        background: rgba(0, 0, 0, 0.7);
-        display: none;
-        justify-content: center;
-        align-items: center;
-        z-index: 1000;
-    }
-
-    .modal-content {
-        background: #fff;
-        padding: 20px;
-        border-radius: 8px;
-        width: 400px;
-        position: relative;
-        margin: 0 auto;
-        top: 200px;
-    }
-
-    .modal-header {
-        border-bottom: 1px solid #444;
-        padding-bottom: 10px;
-        margin-bottom: 15px;
-    }
-    .modal-header > h3{
-        margin: 0 auto;
-    }
-    .modal-close {
-        position: absolute;
-        left: 15px;
-        top: 15px;
-        cursor: pointer;
-        color: #666;
-    }
-
-    .input-group {
-        margin-bottom: 15px;
-    }
-
-    .modal-input {
-        width: 100%;
-        padding: 8px;
-        background: #fff;
-        border: 1px solid #444;
-        color: #5aefcb;
-        border-radius: 7px;
-        margin-top: 5px;
-    }
-
-    .modal-actions {
-        display: flex;
-        gap: 10px;
-        justify-content: left;
-        margin-top: 20px;
-    }
-
-    .confirm-btn {
-        background: #4CAF50;
-        color: white;
-    }
-
-    .cancel-btn {
-        background: #666;
-        color: white;
-    }
-    .action-buttons {
-        margin-top: 20px;
-        display: flex;
-        gap: 10px;
-    }
-
-</style>
-
+@section('content')
 <div class="content-section" id="CategorySection">
     <div class="main-content">
         <div class="myCategory-box">
@@ -164,6 +23,9 @@
                     </div>
                 </div>
             </div>
+
+
+
             <div class="row">
                 <table class="Category-table">
                     <thead>
@@ -328,3 +190,4 @@
         });
     });
 </script>
+@endsection
