+++ resources/views/myTasks/create.blade.php
@@ -1,0 +1,11 @@
+<div class="modal-overlay" id="addModal">
+    <div class="modal-content">
+        <h3>🎯 افزودن تسک جدید</h3>
+        <input type="text" class="modal-input" id="taskTitle" placeholder="عنوان تسک">
+        <textarea class="modal-input" id="taskDescription" placeholder="توضیحات"></textarea>
+        <div class="action-buttons">
+            <button class="btn confirm-btn" id="saveTask">💾 ذخیره</button>
+            <button class="btn cancel-btn" id="cancelAdd">❌ لغو</button>
+        </div>
+    </div>
+</div>
